Move: login, update log 디렉토리 user로 이동
* 관련된 파일 응집시키기위해 이동
