+++ src/modules/user/entity/user.entity.ts
@@ -2,8 +2,8 @@ import { Entity, Column, OneToMany, Generated } from "typeorm";
 import { Bookmark } from "../../bookmark/entity/bookmark.entity";
 import { Comment } from "../../comment/entity/comment.entity";
 import { Like } from "../../like/entity/like.entity";
-import { LoginLog } from "../../login-log/entity/login-log.entity";
-import { UpdateLog } from "../../update-log/entity/update-log.entity";
+import { LoginLog } from "../login-log/entity/login-log.entity";
+import { UpdateLog } from "../update-log/entity/update-log.entity";
 import { Notification } from "../../notifications/entity/notification.entity";
 import { Post } from "../../post/entity/post.entity";
 import { Report } from "../../report/entity/report.entity";
