+++ src/modules/user/login-log/interfaces/ILogin-log.service.ts
@@ -1,4 +1,4 @@
-import { User } from "../../user/entity/user.entity";
+import { User } from "../../entity/user.entity";
 
 export interface ILoginLogService {
   record(user: User, userAgent: string, ip: string): Promise<void>;
