+++ src/modules/user/login-log/login-log.service.ts
@@ -1,7 +1,7 @@
 import { inject, injectable } from "inversify";
-import { TYPES } from "../../core/types.core";
-import { Logger } from "../../shared/utils/logger.util";
-import { User } from "../user/entity/user.entity";
+import { TYPES } from "../../../core/types.core";
+import { Logger } from "../../../shared/utils/logger.util";
+import { User } from "../entity/user.entity";
 import { LoginLog } from "./entity/login-log.entity";
 import { ILoginLogRepository } from "./interfaces/ILogin-log.repository";
 import { ILoginLogService } from "./interfaces/ILogin-log.service";
