+++ src/modules/user/update-log/update-log.module.ts
@@ -1,5 +1,5 @@
 import { ContainerModule } from "inversify";
-import { TYPES } from "../../core/types.core";
+import { TYPES } from "../../../core/types.core";
 import { IUpdateLogRepository } from "./interfaces/IUpdate-log.repository";
 import { IUpdateLogService } from "./interfaces/IUpdate-log.service";
 import { UdpateLogRepository } from "./update-log.repository";
